+++ Engine/Models/GameObjects/IGameObject.cs
@@ -10,8 +10,8 @@ namespace Engine.Models.GameObjects
     public interface IGameObject
     {
         public Vector2 Position { get; set; }
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public float Width { get; set; }
+        public float Height { get; set; }
         public IGraphicsComponent GraphicsComponent {get;set;}
         public void Update(IScene logicContext);
     }
