+++ Engine/Models/GameObjects/LivingEntity.cs
@@ -14,17 +14,17 @@ namespace Engine.Models.GameObjects
     public class LivingEntity : IGameObject
     {
         private Vector2 _position;
-        private double _width;
-        private double _height;
+        private float _width;
+        private float _height;
         private IGraphicsComponent _graphicsComponent;
-        public double Width { get => _width; set => _width = value; }
-        public double Height { get => _height; set => _height = value; }
+        public float Width { get => _width; set => _width = value; }
+        public float Height { get => _height; set => _height = value; }
 
         public IGraphicsComponent GraphicsComponent { get => _graphicsComponent; set => _graphicsComponent = value; }
         public IGameComponent PlayerMovementComponent { get; set; }
         public Vector2 Position { get => _position; set => _position = value; }
 
-        public LivingEntity(IGraphicsComponent graphicsComponent, IGameComponent playerMovementComponent, double width, double height, Vector2 initialPosition, float baseVelocity)
+        public LivingEntity(IGraphicsComponent graphicsComponent, IGameComponent playerMovementComponent, float width, float height, Vector2 initialPosition, float baseVelocity)
         {
             GraphicsComponent = graphicsComponent;
             PlayerMovementComponent = playerMovementComponent;
